Revert "Use EFI_GUID instead of array of bytes in APPLE_PARTITION_INFO_PROTOCOL structure"
This reverts commit 0924ee39aa6cb2b34e79873d7fb639d018239fe5.
APPLE_PARTITION_INFO_PROTOCOL is an Apple-specific extension to HARDDRIVE_DEVICE_PATH, which is declared in EDK2:
https://github.com/tianocore/edk2/blob/7367cc6c24d01b400d2370ffd58ae02854a56b32/MdePkg/Include/Protocol/DevicePath.h#L979
It is best to unify the logic between EDK2 code and our own code in PartitionDxe to avoid misunderstandings.
This is unfortunate, as under normal conditions the value should be declared as a union of UINT8[16], UINT32, and EFI_GUID.
When the first is some default, and the second and third correspond to MBR and GPT formats.